refactor(geometry,core): split oversized modules under ratchet budgets and add adversarial malformed-input tests
Module-size ratchet fixes (split, never raise budgets):
- fast_parse.rs / schema_gen.rs: inline test modules moved to sibling
  *_tests.rs files (ratchet-exempt); fast_parse.rs drops below 400 lines
  so its allowlist row is removed
- boolean/mod.rs: cut-skip heuristics (quality_skips_small_cuts,
  fast_cut_skip_ratio, cutter_below_skip_ratio,
  plane_is_coincident_with_host_face) extracted to
  boolean/cut_heuristics.rs with pub(super) visibility;
  chain_cycle_tests moved to a child file
- voids/synthesis.rs: tests moved to synthesis_tests.rs

Adversarial regression tests for the hardening fixes:
- cycle guard: 2-cycle via root, interior self-loop, 1000-deep legit
  chain (must still walk to the bottom), dangling and null FirstOperand,
  full process() on a cyclic boolean terminates
- total_cmp: all-NaN extents are panic-free and deterministic; pinned
  total-order semantics (-0.0 < 0.0, finite < NaN, last-max tie-break)
- index saturation: u32::MAX / 2^32 / huge-digit-string boundaries,
  i64::MIN and i64::MAX faces, in-range parity with the old wrapping
  parser, whole-polygon drop when any single index is unresolvable,
  checked-multiply threshold indices, valid-index output unchanged
