+++ rust/core/src/fast_parse.rs
@@ -368,96 +368,5 @@ pub fn extract_entity_refs_from_list(bytes: &[u8]) -> Vec<u32> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_parse_coordinates_direct() {
-        let bytes = b"((0.,0.,150.),(0.,40.,140.),(100.,0.,0.))";
-        let coords = parse_coordinates_direct(bytes);
-
-        assert_eq!(coords.len(), 9);
-        assert!((coords[0] - 0.0).abs() < 0.001);
-        assert!((coords[1] - 0.0).abs() < 0.001);
-        assert!((coords[2] - 150.0).abs() < 0.001);
-        assert!((coords[3] - 0.0).abs() < 0.001);
-        assert!((coords[4] - 40.0).abs() < 0.001);
-        assert!((coords[5] - 140.0).abs() < 0.001);
-    }
-
-    #[test]
-    fn test_parse_indices_direct() {
-        let bytes = b"((1,2,3),(2,1,4),(5,6,7))";
-        let indices = parse_indices_direct(bytes);
-
-        assert_eq!(indices.len(), 9);
-        // Should be 0-based (1-based converted)
-        assert_eq!(indices[0], 0); // 1 -> 0
-        assert_eq!(indices[1], 1); // 2 -> 1
-        assert_eq!(indices[2], 2); // 3 -> 2
-        assert_eq!(indices[3], 1); // 2 -> 1
-        assert_eq!(indices[4], 0); // 1 -> 0
-        assert_eq!(indices[5], 3); // 4 -> 3
-    }
-
-    #[test]
-    fn test_parse_indices_direct_rejects_out_of_range() {
-        // 4294967297 = 2^32 + 1. Wrapping arithmetic would map it to 1 (→ 0
-        // after the 1-based fixup), silently aliasing a valid-looking vertex.
-        // Checked+saturate turns it into an obviously out-of-range sentinel
-        // (u32::MAX - 1) the downstream bounds checks drop.
-        let bytes = b"((4294967297,1,2))";
-        let indices = parse_indices_direct(bytes);
-
-        assert_eq!(indices.len(), 3);
-        assert_eq!(
-            indices[0],
-            u32::MAX - 1,
-            "overflowing index must saturate, not wrap to a valid vertex"
-        );
-        assert_eq!(indices[1], 0); // 1 -> 0
-        assert_eq!(indices[2], 1); // 2 -> 1
-    }
-
-    #[test]
-    fn test_parse_scientific_notation() {
-        let bytes = b"((1.5E-10,2.0e+5,-3.14))";
-        let coords = parse_coordinates_direct(bytes);
-
-        assert_eq!(coords.len(), 3);
-        assert!((coords[0] - 1.5e-10).abs() < 1e-15);
-        assert!((coords[1] - 2.0e5).abs() < 1.0);
-        assert!((coords[2] - (-std::f32::consts::PI)).abs() < 0.01);
-    }
-
-    #[test]
-    fn test_parse_negative_numbers() {
-        let bytes = b"((-1.0,-2.5,3.0))";
-        let coords = parse_coordinates_direct(bytes);
-
-        assert_eq!(coords.len(), 3);
-        assert!((coords[0] - (-1.0)).abs() < 0.001);
-        assert!((coords[1] - (-2.5)).abs() < 0.001);
-        assert!((coords[2] - 3.0).abs() < 0.001);
-    }
-
-    #[test]
-    fn test_extract_coordinate_list() {
-        let entity = b"#78=IFCCARTESIANPOINTLIST3D(((0.,0.,150.),(100.,0.,0.)));";
-        let coords = extract_coordinate_list_from_entity(entity).unwrap();
-
-        assert_eq!(coords.len(), 6);
-        assert!((coords[0] - 0.0).abs() < 0.001);
-        assert!((coords[2] - 150.0).abs() < 0.001);
-        assert!((coords[3] - 100.0).abs() < 0.001);
-    }
-
-    #[test]
-    fn test_should_use_fast_path() {
-        assert!(should_use_fast_path("IFCCARTESIANPOINTLIST3D"));
-        assert!(should_use_fast_path("IFCTRIANGULATEDFACESET"));
-        assert!(should_use_fast_path("IfcTriangulatedFaceSet"));
-        assert!(!should_use_fast_path("IFCWALL"));
-        assert!(!should_use_fast_path("IFCEXTRUDEDAREASOLID"));
-    }
-}
+#[path = "fast_parse_tests.rs"]
+mod tests;
