+++ rust/geometry/src/processors/boolean/mod.rs
@@ -22,8 +22,12 @@ use super::swept::{RevolvedAreaSolidProcessor, SweptDiskSolidProcessor};
 use super::tessellated::TriangulatedFaceSetProcessor;
 use crate::router::GeometryProcessor;
 
+mod cut_heuristics;
 mod halfspace_cap;
 mod polygonal_prism;
+use cut_heuristics::{
+    cutter_below_skip_ratio, plane_is_coincident_with_host_face, quality_skips_small_cuts,
+};
 use halfspace_cap::cap_half_space_clip;
 
 /// Maximum recursion depth for nested boolean operations.
@@ -816,95 +820,6 @@ impl BooleanClippingProcessor {
     }
 }
 
-/// Whether this tessellation tier drops sub-threshold boolean cuts (preview
-/// tiers only). `Medium` (the default) and finer keep every cut, so their
-/// geometry is byte-identical to before this optimization.
-fn quality_skips_small_cuts(quality: TessellationQuality) -> bool {
-    matches!(quality, TessellationQuality::Lowest | TessellationQuality::Low)
-}
-
-/// Skip ratio for preview-mode small-cut dropping: cutter max-dimension as a
-/// fraction of host max-dimension. Default 0.10 (≈ Manifold-era load times on
-/// the steel corpus with no visible change to members). Native callers can tune
-/// it via `IFC_LITE_FAST_CUT_RATIO`; in wasm (no env) the default applies.
-fn fast_cut_skip_ratio() -> f64 {
-    use std::sync::OnceLock;
-    static R: OnceLock<f64> = OnceLock::new();
-    *R.get_or_init(|| {
-        std::env::var("IFC_LITE_FAST_CUT_RATIO")
-            .ok()
-            .and_then(|v| v.parse::<f64>().ok())
-            .filter(|v| v.is_finite() && *v > 0.0)
-            .unwrap_or(0.10)
-    })
-}
-
-/// True when `cutter`'s largest bounding-box dimension is below
-/// [`fast_cut_skip_ratio`] of `host`'s — i.e. a small local cut worth skipping in
-/// the preview tiers. Degenerate (zero-extent) hosts never skip.
-fn cutter_below_skip_ratio(host: &Mesh, cutter: &Mesh) -> bool {
-    let max_dim = |m: &Mesh| -> f64 {
-        let (mn, mx) = m.bounds();
-        (((mx.x - mn.x) as f64).max((mx.y - mn.y) as f64)).max((mx.z - mn.z) as f64)
-    };
-    let h = max_dim(host);
-    if h <= 0.0 {
-        return false;
-    }
-    max_dim(cutter) / h < fast_cut_skip_ratio()
-}
-
-/// Decide whether `plane` (point + outward normal) is coincident with one
-/// of the host mesh's axis-aligned bounding-box faces. The check tolerates
-/// numerical noise scaled to the host's diagonal so it works for both
-/// metre-scale residential walls and millimetre-scale connector hardware.
-fn plane_is_coincident_with_host_face(
-    host: &Mesh,
-    plane_point: Point3<f64>,
-    plane_normal: Vector3<f64>,
-) -> bool {
-    let (mn, mx) = host.bounds();
-    let host_min = Point3::new(mn.x as f64, mn.y as f64, mn.z as f64);
-    let host_max = Point3::new(mx.x as f64, mx.y as f64, mx.z as f64);
-    let dx = host_max.x - host_min.x;
-    let dy = host_max.y - host_min.y;
-    let dz = host_max.z - host_min.z;
-    let diag = (dx * dx + dy * dy + dz * dz).sqrt();
-    if diag <= 0.0 {
-        return false;
-    }
-    // 0.1 % of host diagonal, but never less than 1 mm. A 4 m wall ⇒ 4 mm;
-    // a 20 mm fastener ⇒ 1 mm. Tight enough to reject planes that are
-    // unambiguously *outside* the host (the "intentional engulf" case)
-    // while still catching the Revit top-trim that lands exactly on the
-    // wall's top face within float-precision noise.
-    let tol = (diag * 0.001).max(0.001);
-
-    // Test all 8 bbox corners against the plane. If ANY corner is within
-    // `tol` of the plane, the plane is touching (or near-coincident with)
-    // a face. This catches axis-aligned faces (4 corners hit), as well as
-    // edges (2 corners hit) and even single-vertex grazes — all of which
-    // signal that the cut author meant the plane to ride the host surface,
-    // not engulf the body from far away.
-    let corners = [
-        Point3::new(host_min.x, host_min.y, host_min.z),
-        Point3::new(host_max.x, host_min.y, host_min.z),
-        Point3::new(host_min.x, host_max.y, host_min.z),
-        Point3::new(host_max.x, host_max.y, host_min.z),
-        Point3::new(host_min.x, host_min.y, host_max.z),
-        Point3::new(host_max.x, host_min.y, host_max.z),
-        Point3::new(host_min.x, host_max.y, host_max.z),
-        Point3::new(host_max.x, host_max.y, host_max.z),
-    ];
-    for c in &corners {
-        let signed = (c - plane_point).dot(&plane_normal);
-        if signed.abs() <= tol {
-            return true;
-        }
-    }
-    false
-}
-
 impl GeometryProcessor for BooleanClippingProcessor {
     fn process(
         &self,
@@ -931,55 +846,4 @@ impl Default for BooleanClippingProcessor {
 mod halfspace_cap_tests;
 
 #[cfg(test)]
-mod chain_cycle_tests {
-    use super::*;
-
-    /// A self-referential clipping result: `#10`'s FirstOperand is `#10` again,
-    /// with `#20` an `IfcPolygonalBoundedHalfSpace` cutter. Before the visited-id
-    /// guard, `collect_polygonal_chain` walked `current = first` forever.
-    const CYCLIC_IFC: &str = r#"ISO-10303-21;
-HEADER;
-FILE_DESCRIPTION((''),'2;1');
-FILE_NAME('t.ifc','2024-01-01T00:00:00',(''),(''),'','','');
-FILE_SCHEMA(('IFC4'));
-ENDSEC;
-DATA;
-#10=IFCBOOLEANCLIPPINGRESULT(.DIFFERENCE.,#10,#20);
-#20=IFCPOLYGONALBOUNDEDHALFSPACE($,$,$,$);
-ENDSEC;
-END-ISO-10303-21;
-"#;
-
-    #[test]
-    fn collect_polygonal_chain_terminates_on_cyclic_first_operand() {
-        // Run in a worker thread so a regression (infinite loop + unbounded
-        // `chain.push`) is observed as a timeout instead of hanging the suite.
-        let (tx, rx) = std::sync::mpsc::channel();
-        let handle = std::thread::spawn(move || {
-            let content = CYCLIC_IFC.to_string();
-            let mut decoder = EntityDecoder::new(&content);
-            let entity = decoder.decode_by_id(10).expect("decode #10");
-            let processor = BooleanClippingProcessor::new();
-            let result = processor.collect_polygonal_chain(entity, &mut decoder);
-            let _ = tx.send(result.map(|(base, cutters)| (base.id, cutters.len())));
-        });
-
-        let outcome = rx.recv_timeout(std::time::Duration::from_secs(5));
-        assert!(
-            outcome.is_ok(),
-            "collect_polygonal_chain hung on a cyclic FirstOperand chain"
-        );
-        let _ = handle.join();
-
-        let (base_id, cutter_count) = outcome
-            .unwrap()
-            .expect("collect_polygonal_chain returned Err");
-        // The walk bottoms out on the repeated entity and collects the single
-        // PBHS cutter it saw before detecting the cycle.
-        assert_eq!(base_id, 10, "cycle should bottom out on the repeated entity");
-        assert_eq!(
-            cutter_count, 1,
-            "exactly one PBHS cutter collected before the cycle breaks"
-        );
-    }
-}
+mod chain_cycle_tests;
