+++ rust/geometry/src/router/voids/synthesis.rs
@@ -969,31 +969,5 @@ impl GeometryRouter {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// Non-finite file coords (e.g. `1.E999` → +inf) make the bbox-fallback's
-    /// axis extents `inf - inf = NaN`; the old `partial_cmp().unwrap()` panicked
-    /// on that NaN. A zero `axis_dir` forces that fallback branch.
-    #[test]
-    fn remove_internal_membrane_no_panic_on_non_finite_coords() {
-        let mut m = Mesh::new();
-        // 4 triangles (the minimum the membrane pass processes), all x = +inf so
-        // the fallback's ext[0] = inf - inf = NaN reaches the axis-length sort.
-        for t in 0..4u32 {
-            let base = t * 3;
-            for k in 0..3u32 {
-                m.positions
-                    .extend_from_slice(&[f32::INFINITY, t as f32 + k as f32, k as f32]);
-                m.normals.extend_from_slice(&[0.0, 0.0, 1.0]);
-            }
-            m.indices.extend_from_slice(&[base, base + 1, base + 2]);
-        }
-
-        // Zero axis_dir → bbox fallback that sorts the NaN-bearing extents.
-        let out =
-            GeometryRouter::remove_internal_membrane(&m, Vector3::new(0.0, 0.0, 0.0));
-        // Reaching here at all means no panic; sanity-check a well-formed result.
-        assert_eq!(out.indices.len() % 3, 0);
-    }
-}
+#[path = "synthesis_tests.rs"]
+mod tests;
